+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.2.0
+* Future screens are unmounted now when the current screen don't share the pathname with the preview one.
+* Adds routePath to the stack item.
+* Fixes merge of routes.
+* Adds tests.
+* First version of the readme.
+* Adds build.
+
 # v0.1.1
 * Fixes router exports